requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.30.0
openai-whisper>=20231117
pdfplumber>=0.10.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
